+++ src/main/java/allocationproctime/algorithms/FCFS.java
@@ -1,29 +1,23 @@
 package allocationproctime.algorithms;
 
-import java.io.IOException;
-import java.io.Reader;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.TimeUnit;
-
-import com.opencsv.CSVReader;
 
 import allocationproctime.processes.Process;
 import propertieshandler.PropertiesHandler;
 
 /**
  * This class runs simulation for FCFS schedulding algorithm
- * @author amaruszc
+ * @author Arkadiusz Maruszczak
+ * @see BaseAllocationAlgorithm
  *
  */
 public class FCFS extends BaseAllocationAlgorithm {
 
     /**
-     * The constructor.
+     * The constructor for FCFS class.
+     * @param pathToSourceFile  specifies path to source file
+     * @param isUnderTest       is set to true if method is used in tests. As default it is false.
      *
-     * @param pathToSourceFile
      */
     public FCFS(String pathToSourceFile, boolean... isUnderTest) {
         boolean isTest = isUnderTest.length > 0 ? isUnderTest[0] : false;
@@ -39,7 +33,7 @@ public class FCFS extends BaseAllocationAlgorithm {
 
     /**
      * <p>
-     * Begin schedulding processes
+     * Begins schedulding processes
      * </p>
      *
      */
@@ -51,13 +45,13 @@ public class FCFS extends BaseAllocationAlgorithm {
     }
 
     /**
-     *
+     * <p>Implements method from super class.</p>
      */
     protected void executeProcesses() {
 
         boolean allReady = false;
         int i = 0;
-        int waitingTime = 0;
+        double waitingTime = 0;
 
         while (!allReady) {
             waitingTime += this.waitingQueue.get(i).getBurstTime();
