+++ src/main/java/allocationproctime/algorithms/LCFS.java
@@ -1,27 +1,23 @@
 package allocationproctime.algorithms;
 
-import java.io.IOException;
-import java.io.Reader;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.TimeUnit;
-
-import com.opencsv.CSVReader;
 
 import allocationproctime.processes.Process;
 import propertieshandler.PropertiesHandler;
 
+
 /**
  * This class runs simulation for LCFS schedulding algorithm
- * @author amaruszc
+ * @author Arkadiusz Maruszczak
+ * @see BaseAllocationAlgorithm
  *
  */
 public class LCFS extends BaseAllocationAlgorithm{
 
   /**
-   * The constructor.
+   * The constructor for LCFS class.
+   * @param pathToSourceFile  specifies path to source file
+   * @param isUnderTest       is set to true if method is used in tests. As default it is false.
    *
    */
   public LCFS(String pathToSourceFile, boolean... isUnderTest) {
@@ -39,7 +35,7 @@ public class LCFS extends BaseAllocationAlgorithm{
 
   /**
    * <p>
-   * Begin schedulding processes
+   * Begins schedulding processes
    * </p>
    *
    */
@@ -51,13 +47,13 @@ public class LCFS extends BaseAllocationAlgorithm{
   }
 
   /**
-   *
+   *  <p>Implements method from super class</p>
    */
   protected void executeProcesses() {
 
     boolean allReady = false;
     int i = amnt - 1;
-    int waitingTime = 0;
+    double waitingTime = 0;
 
     while (!allReady) {
       waitingTime += this.waitingQueue.get(i).getBurstTime();
